+++ src/main/java/org/manuel/mysportfolio/MysportfolioApplication.java
@@ -1,5 +1,6 @@
 package org.manuel.mysportfolio;
 
+import io.github.manuelarte.spring.queryparameter.mongo.EnableQueryParameter;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.EnableAspectJAutoProxy;
@@ -8,6 +9,7 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 @SpringBootApplication
 @EnableAspectJAutoProxy
 @EnableScheduling
+@EnableQueryParameter
 public class MysportfolioApplication {
 
   public static void main(String[] args) {