+++ src/main/java/org/manuel/mysportfolio/controllers/query/MatchQueryController.java
@@ -1,27 +1,26 @@
 package org.manuel.mysportfolio.controllers.query;
 
+import io.github.manuelarte.spring.queryparameter.mongo.QueryParameter;
 import java.util.Optional;
 import org.bson.types.ObjectId;
 import org.manuel.mysportfolio.config.UserIdProvider;
 import org.manuel.mysportfolio.exceptions.EntityNotFoundException;
-import org.manuel.mysportfolio.model.QueryCriteria;
 import org.manuel.mysportfolio.model.dtos.match.MatchDto;
 import org.manuel.mysportfolio.model.dtos.team.TeamTypeDto;
 import org.manuel.mysportfolio.model.entities.match.Match;
 import org.manuel.mysportfolio.model.entities.match.TeamType;
 import org.manuel.mysportfolio.services.query.MatchQueryService;
-import org.manuel.mysportfolio.transformers.QueryCriteriaToMongoQueryTransformer;
 import org.manuel.mysportfolio.transformers.match.MatchToMatchDtoTransformer;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.data.mongodb.core.query.Query;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -34,17 +33,13 @@ public class MatchQueryController {
   private final MatchToMatchDtoTransformer matchToMatchDtoTransformer;
   private final UserIdProvider userIdProvider;
 
-  private final QueryCriteriaToMongoQueryTransformer queryCriteriaToMongoQueryTransformer;
-
   @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
   public ResponseEntity<Page<MatchDto<TeamTypeDto, TeamTypeDto>>> findByPage(
       @PageableDefault(sort = "startDate", direction = Sort.Direction.DESC) final Pageable pageable,
-      @RequestParam(required = false) Optional<QueryCriteria> q) {
+      @QueryParameter(document = Match.class) final Optional<Query> query) {
     final Page<Match<TeamType, TeamType>> page;
-    if (q.isPresent()) {
-      page = matchQueryService
-          .findQueryAllCreatedBy(queryCriteriaToMongoQueryTransformer.apply(q.get(), Match.class),
-              pageable, getUserId());
+    if (query.isPresent()) {
+      page = matchQueryService.findAllBy(query.get(), pageable, getUserId());
     } else {
       page = matchQueryService.findAllCreatedBy(pageable, getUserId());
     }